refactor(employee-projects): consolidate EmployeeProjectsCell logic and remove container component
