+++ features/employee/index.ts
@@ -1,3 +1,3 @@
 export { EmployeeCard } from './employee-card';
 export { CurrentProjectsCard } from './current-projects-card';
-export { EmployeeProjectsCellContainer } from './employee-projects-cell-container';
+export { EmployeeProjectsCell } from './employee-projects-cell';
